Totally rework how intializing an FsConfig works
The model of a failing new was badwrong for the actual usage of new in
the main function.

Now, instead of possibly failing to initialize, we make an attempt to
read from the current_shell file, but if it's empty we just store a
None for current_shell.

I also moved the actual impls of read_shell and config_path into pure
functions, and made the inherent impl versions pass the correct
arguments to those.